--- bigint/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "sputnikvm-bigint"
-version = "0.3.0"
-license = "Apache-2.0"
-authors = ["Wei Tang <hi@that.world>"]
-description = "Big integer (256-bit and 512-bit) implementation for SputnikVM."
-repository = "https://github.com/ethereumproject/sputnikvm"
-
-[dependencies]
-sputnikvm-rlp = { path = '../rlp' }
