+++ regtests/Cargo.toml
@@ -10,7 +10,6 @@ name = "regtests"
 
 [dependencies]
 sputnikvm = { path = '../sputnikvm' }
-sputnikvm-bigint = { path = '../bigint' }
 gethrpc = { path = '../gethrpc' }
 clap = "2.22"
 serde_json = "0.9"
