--- rlp/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-description = "Recursive-length prefix encoding, decoding, and compression"
-repository = "https://github.com/ethereumproject/sputnikvm"
-license = "MIT/Apache-2.0"
-name = "sputnikvm-rlp"
-version = "0.3.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-
-[dependencies]
-elastic-array = "0.7.0"
-lazy_static = "0.2"
-rustc-serialize = "0.3"
-byteorder = "1.0"
