+++ sputnikvm/Cargo.toml
@@ -17,5 +17,6 @@ ripemd160 = "0.5.2"
 sha2 = "0.5.3"
 digest = { version = "0.5", features = ["std"]}
 secp256k1 = "0.6.2"
-sputnikvm-rlp = { path = '../rlp' }
-sputnikvm-bigint = { path = '../bigint' }
+etcommon-rlp = "0.1.0"
+etcommon-bigint = "0.1.7"
+etcommon-util = "0.1.0"
