+++ sputnikvm/src/util/opcode.rs
